+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## 5.2.0
 
 - Added `WebMessageChannel` and `WebMessageListener` features
-- `AndroidInAppWebViewController.getCurrentWebViewPackage` is available now starting from Android API 21+.
+- Added `canScrollVertically` and `canScrollHorizontally` webview methods
+- `AndroidInAppWebViewController.getCurrentWebViewPackage` is available now starting from Android API 21+
 - Updated Android Gradle distributionUrl version to `5.6.4`
 - Attempt to fix "InAppBrowserActivity.onCreate NullPointerException - Attempt to invoke virtual method 'java.lang.String android.os.Bundle.getString(java.lang.String)' on a null object reference" [#665](https://github.com/pichillilorenzo/flutter_inappwebview/issues/665)
 - Fixed "[iOS] Application crashes when processing onCreateWindow" [#579](https://github.com/pichillilorenzo/flutter_inappwebview/issues/579)
 - Fixed wrong mapping of `NavigationAction` class on Android for `androidHasGesture` and `androidIsRedirect` properties
+- Fixed "Pull to refresh creating problem in some webpages on Android" [#719](https://github.com/pichillilorenzo/flutter_inappwebview/issues/719)
+- Fixed iOS sometimes `scrollView.contentSize` doesn't fit all the `frame.size` available
 
 ## 5.1.0+4